*: fix color log outputs
