+++ tests/runner/go.mod
@@ -11,7 +11,7 @@ require (
 	github.com/ava-labs/avalanche-network-runner v0.0.0-20211220205928-4a40af6d2f7e
 	github.com/ava-labs/avalanchego v1.7.2
 	github.com/ava-labs/quarkvm v0.0.0-00010101000000-000000000000
-	github.com/fatih/color v1.13.0
+	github.com/onsi/ginkgo/v2 v2.0.0
 	github.com/spf13/cobra v1.2.1
 )
 
@@ -30,6 +30,7 @@ require (
 	github.com/deckarep/golang-set v1.7.1 // indirect
 	github.com/decred/dcrd/dcrec/secp256k1/v3 v3.0.0-20200627015759-01fd2de07837 // indirect
 	github.com/ethereum/go-ethereum v1.10.12 // indirect
+	github.com/fatih/color v1.13.0 // indirect
 	github.com/felixge/httpsnoop v1.0.1 // indirect
 	github.com/fjl/memsize v0.0.0-20190710130421-bcb5799ab5e5 // indirect
 	github.com/fsnotify/fsnotify v1.5.1 // indirect
